gravitic-macro/pyproject.toml
[project]
name = "gravitic-macro"
version = "0.1.0"
description = "Macro Risk Signals via Prediction Markets (Polymarket)"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
